+++ elegy/losses/binary_crossentropy.py
@@ -9,6 +9,9 @@ from elegy.losses.loss import Loss, Reduction
 def binary_crossentropy(
     y_true: jnp.ndarray, y_pred: jnp.ndarray, from_logits: bool = False
 ) -> jnp.ndarray:
+    assert abs(y_pred.ndim - y_true.ndim) <= 1
+
+    y_true, y_pred = utils.maybe_expand_dims(y_true, y_pred)
 
     if from_logits:
         return -jnp.mean(y_true * y_pred - jnp.logaddexp(0.0, y_pred), axis=-1)
@@ -62,7 +65,7 @@ class BinaryCrossentropy(Loss):
         module_fn,
         loss=elegy.losses.BinaryCrossentropy(),
         metrics=elegy.metrics.Accuracy(),
-        optimizer=optix.adam(1e-3),
+        optimizer=optax.adam(1e-3),
     )
     ```
     """