+++ elegy/metrics/sparse_categorical_accuracy.py
@@ -54,7 +54,7 @@ class SparseCategoricalAccuracy(Mean):
         module_fn,
         loss=elegy.losses.CategoricalCrossentropy(),
         metrics=elegy.metrics.SparseCategoricalAccuracy(),
-        optimizer=optix.adam(1e-3),
+        optimizer=optax.adam(1e-3),
     )
     ```
     """